+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.51](https://github.com/realityforge/proton/tree/v0.51) (2020-05-28) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.50...v0.51)
 
 Changes in this release: