+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.17`.
 * Upgrade the `com.squareup` artifact to version `1.13.0`.
+* Add a `AbstractStandardProcessor.debugAnnotationProcessingRootElements(...)` helper method that downstream annotation processors can call to improve debuggability.
 
 ### [v0.51](https://github.com/realityforge/proton/tree/v0.51) (2020-05-28) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.50...v0.51)